requirements.txt
annotated-types==0.7.0
anyio==4.4.0
certifi==2024.6.2
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
packaging==24.1
pydantic==2.7.4
pydantic_core==2.18.4
pydub==0.25.1
python-dotenv==1.0.1
pytube==15.0.0
replicate==0.26.0
sniffio==1.3.1
typing_extensions==4.12.2
